upload

+++ module12/assignment.cpp
@@ -28,7 +28,7 @@ using namespace std::chrono;
 #define DEFAULT_PLATFORM 0
 #define DEFAULT_USE_MAP false
 
-int NUM_BUFFER_ELEMENTS = 16;
+float NUM_BUFFER_ELEMENTS = 16;
 #define NUM_SUBBUFFER_ELEMENTS 2
 
 // Function to check and handle OpenCL errors
